+++ config/bootstrap.php
@@ -8,9 +8,21 @@
  * with this source code in the file LICENSE.
  */
 
+use Cake\Core\Configure;
 use Cake\Routing\DispatcherFactory;
 use Lemon\CakePlugin\MaintenanceMode\Routing\Filter\MaintenanceModeFilter;
 
-defined('MAINTENANCE_CONFIG_FILE') || define('MAINTENANCE_CONFIG_FILE', TMP . 'maintenance.php');
-
 DispatcherFactory::add(MaintenanceModeFilter::class);
+
+$defaults = [
+    'lockFile' => TMP . 'maintenance.php',
+    'viewClass' => 'App\View\AppView',
+    'templatePath' => 'Pages',
+    'templateFile' => 'maintenance',
+    'templateLayout' => 'default',
+    'viewVars' => [],
+];
+
+Configure::write('maintenance_mode', array_merge($defaults, (array) Configure::read('maintenance_mode')));
+
+unset($defaults);