+++ CHANGELOG.md
@@ -20,6 +20,13 @@ All Notable changes to project will be documented in this file
 ### Security
 - Nothing
 
+## 0.2.0 - 2016-07-15
+
+### Changed
+
+- [x] Changed to use config instead of content maintenance's lock file
+- [x] Changed command shell to enable and disable maintenance mode
+
 ## 0.1.0 - 2016-04-28
 
 First release with features: